Implement tagged search

+++ cgrep_test.sh
@@ -3,11 +3,19 @@ PYTHON=/c/Users/812426/AppData/Local/Programs/Python/Python38/python.exe
 HOTSPOT=/c/Users/812426/Export/ojdk/jdk14/src/hotspot
 CWD=`pwd`
 
-rm $CWD/cgrep5_test.log 
+if [ -f $CWD/cgrep5_test.log ]
+then
+  rm $CWD/cgrep5_test.log 
+fi  
 
 cd $HOTSPOT
 
-echo "###### TEST 1" >> $CWD/cgrep5_test.log 
+if [ ! -f $CWD/.tags ]
+then
+  ctags -R --c++-types=+px --extra=+q --excmd=pattern --exclude=Makefile --exclude=.tags -f $CWD/.tags
+fi
+
+echo "###### TEST 1 GREP" >> $CWD/cgrep5_test.log 
 $PYTHON $CWD/cgrep5.py -O $CWD/cgrep5_test.log dlopen
 ret=$?
 if [ $ret -eq 34 ]; then
@@ -16,7 +24,7 @@ else
   echo "TEST FAILED expected 34 got $ret"
 fi  
 
-echo "###### TEST 2" >> $CWD/cgrep5_test.log 
+echo "###### TEST 2 GREP" >> $CWD/cgrep5_test.log 
 $PYTHON $CWD/cgrep5.py -O $CWD/cgrep5_test.log "dlopen_.*\(const"
 ret=$?
 if [ $ret -eq 2 ]; then
@@ -25,4 +33,11 @@ else
   echo "TEST FAILED expected 2 got $ret"
 fi  
 
- 
+echo "###### TEST 3 TAG" >> $CWD/cgrep5_test.log 
+$PYTHON $CWD/cgrep5.py -O $CWD/cgrep5_test.log -t $CWD/.tags e:ACCESS_OK
+ret=$?
+if [ $ret -eq 1 ]; then
+  echo "TEST PASSED"
+else
+  echo "TEST FAILED expected 2 got $ret"
+fi
